+++ src/components/excel-format-guide.jsx
@@ -1,12 +1,73 @@
 import React from 'react';
-import { X, FileSpreadsheet, Upload } from 'lucide-react';
+import { X, FileSpreadsheet, Download, CheckCircle } from 'lucide-react';
 import { Button } from '@/components/ui/button';
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card';
 import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from '@/components/ui/table';
+import toast from "react-hot-toast";
 import { Badge } from '@/components/ui/badge';
+import * as XLSX from 'xlsx';
 
-export default function ExcelFormatGuide({ isOpen, onClose, onDownloadTemplate }) {
+export default function ExcelFormatGuide({ isOpen, onClose }) {
   if (!isOpen) return null;
+
+  const showSuccessToast = (message, options = {}) => {
+    toast(
+      () => (
+        <div className="flex items-center gap-3 p-1">
+          <div className="flex-shrink-0 w-8 h-8 bg-green-100 rounded-full flex items-center justify-center">
+            <CheckCircle className="w-4 h-4 text-green-600" />
+          </div>
+          <div className="flex-1 min-w-0">
+            <p className="text-sm font-medium text-gray-900">{message}</p>
+          </div>
+        </div>
+      ),
+      {
+        duration: options.duration || 4000,
+        position: options.position || "bottom-center",
+        style: {
+          background: "white",
+          border: "1px solid #f3f4f6",
+          borderLeft: "3px solid #10b981",
+          boxShadow:
+            "0 10px 15px -3px rgba(0, 0, 0, 0.1), 0 4px 6px -2px rgba(0, 0, 0, 0.05)",
+          borderRadius: "16px",
+          padding: "8px",
+          minWidth: "320px",
+          maxWidth: "450px",
+        },
+      }
+    );
+  };
+
+  const downloadTemplate = () => {
+    const wb = XLSX.utils.book_new();
+    
+    const data = [
+      ['Student Number', 'First Name', 'Last Name', 'Q1', 'Q2', 'ME', 'Final Exam'],
+      ['', '', '', 10, 20, 100, 100],
+      ['22-4539-123', 'Derrick', 'Estopace', 8, 15, 92, 85],
+      ['22-4539-124', 'Kent', 'Abadiano', 9, 18, 95, 90],
+    ];
+    
+    const ws = XLSX.utils.aoa_to_sheet(data);
+    
+    ws['!cols'] = [
+      { wch: 15 },
+      { wch: 12 }, 
+      { wch: 12 },
+      { wch: 8 },
+      { wch: 8 },
+      { wch: 8 },
+      { wch: 12 },
+    ];
+    
+    XLSX.utils.book_append_sheet(wb, ws, 'Student Grades');
+    
+    XLSX.writeFile(wb, 'grade_template.xlsx');
+
+    showSuccessToast('Template downloaded successfully!', { duration: 2000 });
+  };
 
   return (
     <div className="fixed inset-0 bg-black/50 flex items-center justify-center z-50 p-4">
@@ -28,10 +89,10 @@ export default function ExcelFormatGuide({ isOpen, onClose, onDownloadTemplate }
             <Button
               variant="outline"
               size="sm"
-              className="ml-4 border-green-600 text-green-700 hover:bg-green-500 border-green-500 cursor-pointer"
-              onClick={onDownloadTemplate}
+              className="ml-4 border-green-600 text-green-700 hover:bg-green-500 dark:hover:bg-green-900/20 hover:border-green-500 transition-colors"
+              onClick={downloadTemplate}
             >
-              <Upload className="h-4 w-4 mr-2 rotate-180" />
+              <Download className="h-4 w-4 mr-2" />
               Download Template
             </Button>
           </CardDescription>
